+++ Tests/Runtime/BacktraceStackTraceTests.cs
@@ -279,7 +279,7 @@ namespace Backtrace.Unity.Tests.Runtime
         [UnityTest]
         public IEnumerator TestStackTraceCreation_EmptyStackTrace_ValidStackTraceObject()
         {
-            var backtraceStackTrace = new BacktraceStackTrace(new Exception());
+            var backtraceStackTrace = new BacktraceStackTrace(string.Empty, new Exception());
             Assert.IsTrue(backtraceStackTrace.StackFrames.Count == 0);
             yield return null;
         }
@@ -289,7 +289,7 @@ namespace Backtrace.Unity.Tests.Runtime
         {
             var stackTrace = ConvertStackTraceToString(_anrStackTrace);
             var exception = new BacktraceUnhandledException(string.Empty, stackTrace);
-            var backtraceStackTrace = new BacktraceStackTrace(exception);
+            var backtraceStackTrace = new BacktraceStackTrace(string.Empty, exception);
             for (int i = 0; i < _anrStackTrace.Count; i++)
             {
                 var anrStackFrame = _anrStackTrace.ElementAt(i);
@@ -306,7 +306,7 @@ namespace Backtrace.Unity.Tests.Runtime
         {
             var stackTrace = ConvertStackTraceToString(_mixModeCallStack);
             var exception = new BacktraceUnhandledException(string.Empty, stackTrace);
-            var backtraceStackTrace = new BacktraceStackTrace(exception);
+            var backtraceStackTrace = new BacktraceStackTrace(string.Empty, exception);
             //skip first frame
             int startIndex = exception.Header ? 1 : 0;
             for (int i = startIndex; i < _mixModeCallStack.Count; i++)
@@ -329,7 +329,7 @@ namespace Backtrace.Unity.Tests.Runtime
 
                 string message = "message";
                 var exception = new BacktraceUnhandledException(message, stackTrace);
-                var backtraceStackTrace = new BacktraceStackTrace(exception);
+                var backtraceStackTrace = new BacktraceStackTrace(string.Empty, exception);
 
                 //skip first frame
                 int startIndex = exception.Header ? 1 : 0;